+++ PostHogTests/PostHogContextTest.swift
@@ -74,14 +74,11 @@ class PostHogContextTest: QuickSpec {
             expect(context["$app_version"] as? String) != nil
             expect(context["$app_build"] as? String) != nil
             expect(context["$app_namespace"] as? String) != nil
-            // expect(context["$locale"] as? String) != nil
 
             #if os(iOS) || os(tvOS) || os(visionOS)
                 expect(context["$os_name"] as? String) != nil
                 expect(context["$os_version"] as? String) != nil
                 expect(context["$device_type"] as? String) != nil
-                // expect(context["$device_manufacturer"] as? String) == "Apple"
-                // expect(context["$device_model"] as? String) != nil
             #endif
 
             expect(context["$lib"] as? String) == "posthog-ios"